Split platform dependent parts for file locks screen
