+++ Action.c
@@ -27,16 +27,13 @@ in the source distribution for its full text.
 #include "MainPanel.h"
 #include "OpenFilesScreen.h"
 #include "Process.h"
+#include "ProcessLocksScreen.h"
 #include "ProvideCurses.h"
 #include "ScreenManager.h"
 #include "SignalsPanel.h"
 #include "TraceScreen.h"
 #include "Vector.h"
 #include "XUtils.h"
-
-#ifdef HTOP_LINUX
-#include "linux/ProcessLocksScreen.h"
-#endif
 
 
 Object* Action_pickFromVector(State* st, Panel* list, int x, bool followProcess) {
@@ -375,7 +372,6 @@ static Htop_Reaction actionLsof(State* st) {
    return HTOP_REFRESH | HTOP_REDRAW_BAR;
 }
 
-#ifdef HTOP_LINUX
 static Htop_Reaction actionShowLocks(State* st) {
    Process* p = (Process*) Panel_getSelected(st->panel);
    if (!p) return HTOP_OK;
@@ -386,7 +382,6 @@ static Htop_Reaction actionShowLocks(State* st) {
    CRT_enableDelay();
    return HTOP_REFRESH | HTOP_REDRAW_BAR;
 }
-#endif
 
 static Htop_Reaction actionStrace(State* st) {
    Process* p = (Process*) Panel_getSelected(st->panel);
@@ -452,9 +447,7 @@ static const struct { const char* key; const char* info; } helpRight[] = {
    { .key = "      e: ", .info = "show process environment" },
    { .key = "      i: ", .info = "set IO priority" },
    { .key = "      l: ", .info = "list open files with lsof" },
-#ifdef HTOP_LINUX
    { .key = "      x: ", .info = "list file locks of process" },
-#endif
    { .key = "      s: ", .info = "trace syscalls with strace" },
    { .key = "      w: ", .info = "wrap process command in multiple lines" },
    { .key = " F2 C S: ", .info = "setup" },
@@ -640,9 +633,7 @@ void Action_setBindings(Htop_Action* keys) {
    keys['S'] = actionSetup;
    keys['C'] = actionSetup;
    keys[KEY_F(2)] = actionSetup;
-#ifdef HTOP_LINUX
    keys['x'] = actionShowLocks;
-#endif
    keys['l'] = actionLsof;
    keys['s'] = actionStrace;
    keys[' '] = actionTag;
