+++ freebsd/Platform.c
@@ -243,6 +243,17 @@ char* Platform_getProcessEnv(pid_t pid) {
    return env;
 }
 
+char* Platform_getInodeFilename(pid_t pid, ino_t inode) {
+    (void)pid;
+    (void)inode;
+    return NULL;
+}
+
+FileLocks_ProcessData* Platform_getProcessLocks(pid_t pid) {
+    (void)pid;
+    return NULL;
+}
+
 bool Platform_getDiskIO(DiskIOData* data) {
 
    if (devstat_checkversion(NULL) < 0)
