+++ unsupported/Platform.c
@@ -141,6 +141,17 @@ char* Platform_getProcessEnv(pid_t pid) {
    return NULL;
 }
 
+char* Platform_getInodeFilename(pid_t pid, ino_t inode) {
+    (void)pid;
+    (void)inode;
+    return NULL;
+}
+
+FileLocks_ProcessData* Platform_getProcessLocks(pid_t pid) {
+    (void)pid;
+    return NULL;
+}
+
 bool Platform_getDiskIO(DiskIOData* data) {
    (void)data;
    return false;
